requirements.txt
# Геологический эмбеддер — зависимости

# ─── Подготовка данных ────────────────────────────────────────────────────────
pypdf>=4.0.0              # извлечение текста из PDF
python-docx>=1.1.0        # чтение DOCX
langchain-text-splitters>=0.2.0  # рекурсивное чанкование
rank-bm25>=0.2.2          # BM25 hard negatives
tqdm>=4.66.0              # прогресс-бары

# ─── Обучение ─────────────────────────────────────────────────────────────────
torch>=2.1.0
sentence-transformers>=2.7.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.28.0

# ─── Оценка ───────────────────────────────────────────────────────────────────
numpy>=1.26.0
scikit-learn>=1.4.0

# ─── Сервинг ──────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0

# ─── MLflow (трекинг экспериментов) ──────────────────────────────────────────
mlflow>=2.13.0

# ─── LLM API (для генерации запросов и adversarial негативов) ─────────────────
openai>=1.30.0            # опционально, если используется OpenAI backend
python-dotenv>=1.0.0      # загрузка .env
